* Intermediate: Now uses the W0CHP-modified MMDVMHost binary, to exract DMR first names, and display them on the dashboard rows.   - NOTE: Current Caller feature must be turned on for this to be enabled, plus the country flags.

+++ mmdvmhost/lh.php
@@ -41,8 +41,14 @@ $Flags->LoadFlags();
       <th><a class="tooltip" href="#"><?php echo $lang['time'];?> (<?php echo date('T')?>)<span><b>Time in <?php echo date('T')?> time zone</b></span></a></th>
       <th><a class="tooltip" href="#"><?php echo $lang['mode'];?><span><b>Transmitted Mode</b></span></a></th>
       <th width="85px"><a class="tooltip" href="#"><?php echo $lang['callsign'];?><span><b>Callsign</b></span></a></th>
+<?php
+    if (file_exists("/etc/.CALLERDETAILS")) {
+?>
       <th width="50px"><a class="tooltip" href="#">Country<span><b>Country</b></span></a></th>
       <th><a class="tooltip" href="#">Name<span><b>Name</b></span></a></th>
+<?php
+    }
+?>
       <th><a class="tooltip" href="#"><?php echo $lang['target'];?><span><b>Target, D-Star Reflector, DMR Talk Group etc</b></span></a></th>
       <th><a class="tooltip" href="#"><?php echo $lang['src'];?><span><b>Received from source</b></span></a></th>
       <th><a class="tooltip" href="#"><?php echo $lang['dur'];?>(s)<span><b>Duration in Seconds</b></span></a></th>
@@ -83,15 +89,31 @@ for ($i = 0;  ($i <= $lastHeardRows - 1); $i++) {
 		echo"<td align=\"left\">$local_time</td>";
 		echo "<td align=\"left\">".str_replace('Slot ', 'TS', $listElem[1])."</td>";
 		if (is_numeric($listElem[2]) || strpos($listElem[2], "openSPOT") !== FALSE) {
+		    if (file_exists("/etc/.CALLERDETAILS")) {
                         echo "<td align=\"left\">$listElem[2]</td><td colspan='2'>$flContent</td>";
+		    } else {
+                        echo "<td align=\"left\">$listElem[2]</td>";
+		    }
 		} elseif (!preg_match('/[A-Za-z].*[0-9]|[0-9].*[A-Za-z]/', $listElem[2])) {
+		    if (file_exists("/etc/.CALLERDETAILS")) {
                         echo "<td align=\"left\">$listElem[2]</td><td colspan='2'>$flContent</td>";
+		    } else {
+                        echo "<td align=\"left\">$listElem[2]</td>";
+		    }
 		} else {
 			if (strpos($listElem[2],"-") > 0) { $listElem[2] = substr($listElem[2], 0, strpos($listElem[2],"-")); }
 			if ( $listElem[3] && $listElem[3] != '    ' ) {
+			    if (file_exists("/etc/.CALLERDETAILS")) {
 				echo "<td align=\"left\"><a href=\"http://www.qrz.com/db/$listElem[2]\" target=\"_blank\">$listElem[2]</a>//$listElem[8]</td><td>$flContent</td><td align='left'>$listElem[11]</td>";
+			    } else {
+				echo "<td align=\"left\"><a href=\"http://www.qrz.com/db/$listElem[2]\" target=\"_blank\">$listElem[2]</a>//$listElem[8]</td>";
+			    }
 			} else {
+			    if (file_exists("/etc/.CALLERDETAILS")) {
 				echo "<td align=\"left\"><a href=\"http://www.qrz.com/db/$listElem[2]\" target=\"_blank\">$listElem[2]</a></td><td>$flContent</td><td align='left'>$listElem[11]</td>";
+			    } else {
+				echo "<td align=\"left\"><a href=\"http://www.qrz.com/db/$listElem[2]\" target=\"_blank\">$listElem[2]</a></td></td>";
+			    }
 			}
 		}
 		if (strlen($listElem[4]) == 1) { $listElem[4] = str_pad($listElem[4], 8, " ", STR_PAD_LEFT); }
